Time: 248 ms (76.07%), Space: 135.9 MB (48.49%) - LeetHub

+++ 2090-k-radius-subarray-averages/2090-k-radius-subarray-averages.cpp
@@ -5,11 +5,11 @@ public:
         if(n < 2*k + 1)
             return vector<int>(n, -1);
         
-        long long winSum = 0;
+        long winSum = 0;
         vector<int> v(k, -1);
-        
         for(int i=0; i < (2*k + 1); i++)
             winSum += nums[i];
+        
         for(int i=k; i < n-k; i++) {
             v.push_back(winSum / (2*k + 1));
             winSum -= nums[i - k];
@@ -18,7 +18,6 @@ public:
             else
                 break;
         }
-        
         while(k--)
             v.push_back(-1);
         return v;
